requirements.txt
# ── Flask Web Framework ──
flask

# ── Document Parsing ──
PyMuPDF             
python-docx         

# ── NLP & Skill Extraction ──
spacy               
scikit-learn         

# ── Embeddings & Semantic Similarity ──
sentence-transformers 

# ── ML Model Persistence ──
joblib               

# ── Data Processing ──
pandas
numpy

# ── PyTorch (required by sentence-transformers) ──
torch
